+++ dataworks-public-20200518/core/Models/ListQualityResultsByEntityResponseBody.cs
@@ -152,7 +152,7 @@ namespace AlibabaCloud.SDK.Dataworks_public20200518.Models
                 /// </summary>
                 [NameInMap("EntityId")]
                 [Validation(Required=false)]
-                public int? EntityId { get; set; }
+                public long? EntityId { get; set; }
 
                 /// <summary>
                 /// The ID of the primary key.
@@ -187,7 +187,7 @@ namespace AlibabaCloud.SDK.Dataworks_public20200518.Models
                 /// </summary>
                 [NameInMap("Id")]
                 [Validation(Required=false)]
-                public int? Id { get; set; }
+                public long? Id { get; set; }
 
                 /// <summary>
                 /// Indicates whether the monitoring is discrete monitoring. Valid values:
@@ -301,7 +301,7 @@ namespace AlibabaCloud.SDK.Dataworks_public20200518.Models
                 /// </summary>
                 [NameInMap("RuleId")]
                 [Validation(Required=false)]
-                public int? RuleId { get; set; }
+                public long? RuleId { get; set; }
 
                 /// <summary>
                 /// The time when the monitoring ended.
